requirements.txt
treamlit
joblib
pandas
folium
streamlit-folium
scikit-learn
fastapi
uvicorn

